Fix data-plane robustness and stats-integrity issues
- tx: reject mirror mode combined with UDP debug (-m -D) at startup
  instead of aborting on assert(idx >= 0) on the first packet
- tx: track SO_RXQ_OVFL per input socket (array indexed by socket)
  and only count positive deltas, fixing a uint32 underflow that
  reported ~4.29e9 bogus dropped/incoming packets with >=2 wlans
- tx: write the session TLV length big-endian (htons) to match the
  wire contract and rx get_tag (ntohs)
- rx: widen packet_seq/seq to uint64 so loss accounting and the
  packet-loss callback no longer wrap within a long-lived session
- rx: drop the redundant close(fd) in ~Receiver; pcap_close already
  closes the selectable fd (fixes a double-close)
- wfb_tun: replace asserts on runtime I/O results and inbound packet
  framing with graceful drop/retry paths so a benign EAGAIN, short
  write, EOF, or misframed batch no longer abort()s the tunnel;
  NUL-terminate the interface name copied via strncpy

+++ src/rx.hpp
@@ -39,7 +39,7 @@ class PacketLossListener
 {
 public:
     virtual ~PacketLossListener() = default;
-    virtual void on_packet_loss(uint32_t lost_count, uint32_t last_seq, uint32_t new_seq) = 0;
+    virtual void on_packet_loss(uint32_t lost_count, uint64_t last_seq, uint64_t new_seq) = 0;
 };
 
 typedef enum {
@@ -234,7 +234,7 @@ private:
     int fec_n;  // RS total number of fragments in block
     uint8_t session_hash[crypto_generichash_BYTES];
 
-    uint32_t seq;
+    uint64_t seq;
     rx_ring_item_t rx_ring[RX_RING_SIZE];
     int rx_ring_front; // current packet
     int rx_ring_alloc; // number of allocated entries
